wip: more

+++ apps/shuffle/util/cperf.py
@@ -32,14 +32,14 @@ import time
 import traceback
 import fcntl
 
-# If a server's id appears as a key in this dictionary, it means we
-# have started cp_node running on that node. The value of each entry is
+# If a server's id appears as a key in this dictionary, it means we have
+# started shuffle_node running on that node. The value of each entry is
 # a Popen object that can be used to communicate with the node.
-# active_nodes = {}
-active_nodes = []
+active_nodes = {}
 
-# The range of nodes currently running cp_node servers.
-server_nodes = range(0,0)
+# A dictionary whose keys are server ids and whose values are CloudLab public
+# IP addresses which are used to establish ssh connections.
+node_ssh_addrs = {}
 
 # Directory containing log files.
 log_dir = ''
@@ -48,7 +48,7 @@ log_dir = ''
 log_file = ''
 
 # Open file (in the log directory) where log messages should be written.
-cperf_log_file = 0
+cperf_log = 0
 
 # Indicates whether we should generate additional log messages for debugging
 verbose = False
@@ -69,6 +69,33 @@ default_defaults = {
     'log_dir':             'logs/' + time.strftime('%Y%m%d%H%M%S'),
     'log_file':            'node.log',
 }
+
+
+def log(message):
+    """
+    Write the a log message both to stdout and to the cperf log file.
+
+    message:  The log message to write; a newline will be appended.
+    """
+    global cperf_log
+    print(message)
+    cperf_log.write(message)
+    cperf_log.write("\n")
+
+
+def vlog(message):
+    """
+    Log a message, like log, but if verbose logging isn't enabled, then
+    log only to the cperf log file, not to stdout.
+
+    message:  The log message to write; a newline will be appended.
+    """
+    global cperf_log, verbose
+    if verbose:
+        print(message)
+    cperf_log.write(message)
+    cperf_log.write("\n")
+
 
 def get_parser(description, usage, defaults = {}):
     """
@@ -129,6 +156,8 @@ def get_parser(description, usage, defaults = {}):
             % (defaults['log_file']))
     parser.add_argument('-v', '--verbose', dest='verbose', action='store_true',
             help='Enable verbose output in node logs')
+    parser.add_argument('--plot-only', dest='plot_only', action='store_true',
+            help='Don\'t run experiments; generate plot(s) with existing data')
     return parser
 
 
@@ -136,16 +165,19 @@ def init(options):
     """
     Initialize various global state, such as the log file.
     """
-    global log_dir, log_file, cperf_log_file, verbose
+    global log_dir, log_file, cperf_log, verbose
     log_dir = options.log_dir
     log_file = options.log_file
 
-    if os.path.exists(log_dir):
-        shutil.rmtree(log_dir)
-    os.makedirs(log_dir)
-    os.symlink(log_dir, log_dir + "../latest", target_is_directory=True)
+    if not options.plot_only:
+        if os.path.exists(log_dir):
+            shutil.rmtree(log_dir)
+        os.makedirs(log_dir)
+        latest = log_dir + "/../latest"
+        os.unlink(latest)
+        os.symlink(os.path.abspath(log_dir), latest, target_is_directory=True)
 
-    cperf_log_file = open("%s/cperf.log" % (log_dir), "a")
+    cperf_log = open("%s/cperf.log" % (log_dir), "a")
     verbose = options.verbose
     vlog("cperf starting at %s" % (date_time))
     s = ""
@@ -156,31 +188,28 @@ def init(options):
         s += ("--%s: %s" % (name, str(opts[name])))
     vlog("Options: %s" % (s))
 
+    cluster_init(options.server_list)
 
-def log(message):
+
+def cluster_init(server_list):
     """
-    Write the a log message both to stdout and to the cperf log file.
+    Read the server list file to obtain the list of the nodes in the cluster.
 
-    message:  The log message to write; a newline will be appended.
+    :param server_list:
+        Each line of the file contains a host name, its public IP address, and
+        its private IP address, separated by whitespaces.
+    :return:
+        A list of nodes in the cluster; each node is represented by its hostname
+        and public IP address.
     """
-    global cperf_log_file
-    print(message)
-    cperf_log_file.write(message)
-    cperf_log_file.write("\n")
-
-
-def vlog(message):
-    """
-    Log a message, like log, but if verbose logging isn't enabled, then
-    log only to the cperf log file, not to stdout.
-
-    message:  The log message to write; a newline will be appended.
-    """
-    global cperf_log_file, verbose
-    if verbose:
-        print(message)
-    cperf_log_file.write(message)
-    cperf_log_file.write("\n")
+    global node_ssh_addrs
+    for line in open(server_list):
+        try:
+            hostname, public_ip = line.strip().split(" ")[:2]
+            id = int(hostname[2:])
+            node_ssh_addrs[id] = public_ip
+        except:
+            print("failed to parse '%s' in file %s" % (line, server_list))
 
 
 def wait_output(string, nodes, cmd, time_limit=10.0):
@@ -204,12 +233,12 @@ def wait_output(string, nodes, cmd, time_limit=10.0):
     while time.time() < (start_time + time_limit):
         for id in nodes:
             data = active_nodes[id].stdout.read(1000)
-            if data != None:
+            if data is not None:
                 print_data = data
                 if print_data.endswith(string):
                     print_data = print_data[:(len(data) - len(string))]
                 if print_data != "":
-                    log("output from node-%d: '%s'" % (id, print_data))
+                    log("output from rc%02d: '%s'" % (id, print_data))
                 outputs[id] += data
         bad_node = -1
         for id in nodes:
@@ -219,74 +248,111 @@ def wait_output(string, nodes, cmd, time_limit=10.0):
         if bad_node < 0:
             return
         if (time.time() > (start_time + time_limit)) and not printed:
-            log("expected output from node-%d not yet received "
+            log("expected output from rc%02d not yet received "
             "after command '%s': expecting '%s', got '%s'"
             % (bad_node, cmd, string, outputs[bad_node]))
-            printed = True;
+            printed = True
         time.sleep(0.1)
-    raise Exception("bad output from node-%d after command '%s': "
+    raise Exception("bad output from rc%02d after command '%s': "
             "expected '%s', got '%s'"
             % (bad_node, cmd, string, outputs[bad_node]))
 
 
-def get_cluster_nodes(cluster_config):
+def start_nodes(r, shell_cmd):
     """
-    Read the cluster config file to obtain the list of the nodes in the cluster.
-
-    :param cluster_config:
-        Each line of the file contains a host name, its public IP address, and
-        its private IP address, separated by whitespaces.
-    :return:
-        A list of nodes in the cluster; each node is represented by its hostname
-        and public IP address.
+    Start up shuffle_node on a group of nodes.
+    r:          The range of nodes on which to start shuffle_node, if it isn't
+                already running
+    shell_cmd:  Shell command used to start shuffle_node
     """
-    cluster_nodes = []
-    for line in open(cluster_config):
-        try:
-            hostname, public_ip, private_ip = line.strip().split(" ")
-            cluster_nodes.append((hostname, public_ip))
-        except:
-            print("failed to parse '%s'" % line)
-    return cluster_nodes
-
-
-def cluster_exec(servers, cmd):
-    """
-    Execute the same command on a group of servers.
-
-    :param servers:
-        IP addresses of the servers, in text format.
-    :param command:
-        Command to run.
-    """
-    for hostname, ip in servers:
-        vlog("Executing '%s' on %s" % (cmd, hostname))
-        subprocess.Popen(["ssh", "-o", "StrictHostKeyChecking=no", ip, cmd])
-        active_nodes.append((hostname, ip))
+    global active_nodes
+    started = []
+    for id in r:
+        if id in active_nodes:
+            continue
+        vlog("Starting shuffle_node on rc%02d" % (id))
+        node = subprocess.Popen(["ssh", "-o", "StrictHostKeyChecking=no",
+                node_ssh_addrs[id], shell_cmd], encoding="utf-8",
+                stdin=subprocess.PIPE, stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT)
+        fl = fcntl.fcntl(node.stdin, fcntl.F_GETFL)
+        fcntl.fcntl(node.stdin, fcntl.F_SETFL, fl | os.O_NONBLOCK)
+        fl = fcntl.fcntl(node.stdout, fcntl.F_GETFL)
+        fcntl.fcntl(node.stdout, fcntl.F_SETFL, fl | os.O_NONBLOCK)
+        active_nodes[id] = node
+        started.append(id)
+    # FIXME: what is this? am I supposed to print something in shuffle_node.cc?
+    wait_output("% ", started, "ssh")
+    # FIXME: shall I implement "log" command in shuffle_node? maybe w/o --level option?
+    # log_level = "normal"
+    # if verbose:
+    #     log_level = "verbose"
+    # command = "log --file node.log --level %s" % (log_level)
+    # for id in started:
+    #     active_nodes[id].stdin.write(command + "\n")
+    #     active_nodes[id].stdin.flush()
+    # wait_output("% ", started, command)
 
 
 def stop_nodes():
     """
     Exit all of the nodes that are currently active.
     """
-    # global active_nodes, server_nodes
-    # for id, popen in homa_prios.items():
-    #     subprocess.run(["ssh", "-o", "StrictHostKeyChecking=no",
-    #             "node-%d" % id, "sudo", "pkill", "homa_prio"])
-    #     try:
-    #         popen.wait(5.0)
-    #     except subprocess.TimeoutExpired:
-    #         log("Timeout killing homa_prio on node-%d" % (id))
-    # for node in active_nodes.values():
-    #     node.stdin.write("exit\n")
-    #     try:
-    #         node.stdin.flush()
-    #     except BrokenPipeError:
-    #         log("Broken pipe to node-%d" % (id))
-    # for node in active_nodes.values():
-    #     node.wait(5.0)
-    for hostname, ip in active_nodes:
+    global active_nodes, node_ssh_addrs, log_file
+    for id, node in active_nodes.items():
+        node.stdin.write("exit\n")
+        try:
+            node.stdin.flush()
+        except BrokenPipeError:
+            log("Broken pipe to rc%02d" % id)
+    for node in active_nodes.values():
+        node.wait(5.0)
+    for id in active_nodes:
+        ip = node_ssh_addrs[id]
         subprocess.run(["rsync", "-rtvq", "%s:%s" % (ip, log_file),
-                "%s/%s.log" % (log_dir, hostname)])
+                "%s/rc%02d.log" % (log_dir, id)])
     active_nodes.clear()
-    # server_nodes = range(0,0)
+
+
+def do_cmd(command, r, r2 = range(0,0)):
+    """
+    Execute a shuffle_node command on a given group of nodes.
+    command:    A command to execute on each node
+    r:          A group of node ids on which to run the command (range, list,
+                etc.)
+    r2:         An optional additional group of node ids on which to run the
+                command; if a note is present in both r and r2, the
+                command will only be performed once
+    """
+    global active_nodes
+    nodes = []
+    for id in r:
+        nodes.append(id)
+    for id in r2:
+        if id not in r:
+            nodes.append(id)
+    for id in nodes:
+        vlog("Command for rc%02d: %s" % (id, command))
+        active_nodes[id].stdin.write(command + "\n")
+        try:
+            active_nodes[id].stdin.flush()
+        except BrokenPipeError:
+            log("Broken pipe to rc%02d" % (id))
+    wait_output("% ", nodes, command)
+
+
+def do_ssh(command, nodes):
+    """
+    Use ssh to execute a particular shell command on a group of nodes.
+    command:  command to execute on each node (a list of argument words)
+    nodes:    specifies ids of the nodes on which to execute the command:
+              should be a range, list, or other object that supports "in"
+    """
+    global node_ssh_addrs
+    vlog("ssh command on nodes %s: %s" % (str(nodes), " ".join(command)))
+    for id in nodes:
+        subprocess.run(["ssh", node_ssh_addrs[id]] + command,
+                stdout=subprocess.DEVNULL)
+
+
+# FIXME: run_experiments, scan_log, etc. are dropped for now; do I need them?
